Fixed missing cast

+++ src/metadata.cpp
@@ -37,8 +37,8 @@ void readMetadata(Parameters *params) {
 
   // What is the fraction of particles we will read in unused cells to reduce
   // I/O calls?
-  metadata->gap_fill_fraction =
-      params->getParameter<double>("Input/part_gap_fill_fraction", 0.01);
+  metadata->gap_fill_fraction = static_cast<double>(
+      params->getParameter<double>("Input/part_gap_fill_fraction", 0.01));
 
   message("Reading data from: %s", metadata->input_file.c_str());
